+++ tests/test_session_queue.py
@@ -11,7 +11,8 @@ from invokeai.app.services.session_queue.session_queue_common import (
     prepare_values_to_insert,
 )
 from invokeai.app.services.shared.graph import Graph, GraphExecutionState
-from tests.aa_nodes.test_nodes import PromptTestInvocation
+
+from .test_nodes import PromptTestInvocation
 
 
 @pytest.fixture
